+++ packages/wrapper/src/languageClientWrapper.ts
@@ -6,9 +6,8 @@
 import { BrowserMessageReader, BrowserMessageWriter } from 'vscode-languageserver-protocol/browser.js';
 import { CloseAction, ErrorAction, LanguageClientOptions, MessageTransports, State } from 'vscode-languageclient/browser.js';
 import { MonacoLanguageClient, ConnetionConfigOptions, WorkerConfigOptionsDirect, WorkerConfigOptionsParams } from 'monaco-languageclient';
-import { Logger } from 'monaco-languageclient/tools';
+import { createUrl, Logger } from 'monaco-languageclient/tools';
 import { toSocket, WebSocketMessageReader, WebSocketMessageWriter } from 'vscode-ws-jsonrpc';
-import { createUrl } from './utils.js';
 
 export interface ConnectionConfig {
     options: ConnetionConfigOptions;